+++ app/components/navigation/right.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import connectToStores from 'alt-utils/lib/connectToStores';
 import { Nav, NavItem } from 'react-bootstrap';
 import ApplicationActions from 'actions/application';
-import SessionActions from 'actions/session';
 import session from 'services/session';
 
 @connectToStores
@@ -28,7 +27,7 @@ export default class NavigationRight extends React.Component {
   }
 
   signOut() {
-    SessionActions.delete(session.currentUser());
+    session.delete();
   }
 
   render() {