+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- Extract session actions into session service
+([#40](https://github.com/fs/react-base/pull/40))
 - Add new rules in postcss sorting config
 ([#39](https://github.com/fs/react-base/pull/39))
 - Upgrade Node.js to 6.3.x